backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pyotp==2.9.0
httpx==0.27.2
google-auth==2.35.0
email-validator==2.2.0
